requirements.txt
django==5.2.11
djangorestframework==3.16.1